** DIVERGENCE 1 ** Refactor reading in the S matrix
This highlighted the issue that in the original code, if a compound appears on both sides of the equation, then it has its stochiometry set to its stochiometric number for the right hand side only. This meant that e.g. in R102, where H+_[cyt] appears with a coefficient of 1 on both sides, it ended up with a stochiometry of 1 for this equation rather than 0.